app/requirements.txt
# requirements.txt
fastapi
uvicorn
langchain
openai
python-multipart
cryptography
PyPDF2
streamlit
requests
python-dotenv
langchain-community
pypdf
tiktoken